requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Environment and Config
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytz==2023.3 